requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
alembic=1.7.6=pypi_0
arviz=0.11.4=pyhd8ed1ab_0
attrs=21.4.0=pypi_0
autopage=0.5.0=pypi_0
binutils_impl_linux-64=2.35.1=h27ae35d_9
binutils_linux-64=2.35.1=h454624a_30
blas=1.0=mkl
bottleneck=1.3.2=py39hdd57654_1
bzip2=1.0.8=h7f98852_4
c-ares=1.18.1=h7f8727e_0
ca-certificates=2020.10.14=0
cachetools=5.0.0=pyhd8ed1ab_0
certifi=2021.10.8=py39h06a4308_2
cftime=1.5.1.1=py39hce1f21e_0
cliff=3.10.0=pypi_0
cmaes=0.8.2=pypi_0
cmd2=2.3.3=pypi_0
colorlog=6.6.0=pypi_0
curl=7.80.0=h7f8727e_0
cycler=0.11.0=pyhd8ed1ab_0
dill=0.3.4=pyhd8ed1ab_0
fastprogress=1.0.2=pyhd8ed1ab_0
filelock=3.5.1=pyhd8ed1ab_0
fonttools=4.28.5=pypi_0
freetype=2.10.4=h0708190_1
gcc_impl_linux-64=9.3.0=h6df7d76_17
gcc_linux-64=9.3.0=h1ee779e_30
greenlet=1.1.2=pypi_0
gxx_impl_linux-64=9.3.0=hbdd7822_17
gxx_linux-64=9.3.0=h7e70986_30
hdf4=4.2.13=h3ca952b_2
hdf5=1.10.6=nompi_h6a2412b_1114
importlib-metadata=4.11.1=py39hf3d152e_0
importlib_metadata=4.11.1=hd8ed1ab_0
intel-openmp=2021.4.0=h06a4308_3561
joblib=1.1.0=pyhd8ed1ab_0
jpeg=9d=h7f8727e_0
kernel-headers_linux-64=2.6.32=he073ed8_15
kiwisolver=1.3.2=py39h295c915_0
krb5=1.19.2=hcc1bbae_0
lcms2=2.12=hddcbb42_0
ld_impl_linux-64=2.35.1=h7274673_9
libblas=3.9.0=12_linux64_mkl
libcblas=3.9.0=12_linux64_mkl
libcurl=7.80.0=h0b77cf5_0
libedit=3.1.20191231=he28a2e2_2
libev=4.33=h516909a_1
libffi=3.3=he6710b0_2
libgcc-devel_linux-64=9.3.0=hb95220a_17
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=11.2.0=h69a702a_12
libgfortran5=11.2.0=h5c6108e_12
libgomp=9.3.0=h5101ec6_17
liblapack=3.9.0=12_linux64_mkl
libnetcdf=4.8.1=h42ceab0_1
libnghttp2=1.46.0=hce63b2e_0
libpng=1.6.37=h21135ba_2
libssh2=1.9.0=h1ba5d50_1
libstdcxx-devel_linux-64=9.3.0=hf0c5c8d_17
libstdcxx-ng=9.3.0=hd4cf53a_17
libtiff=4.2.0=h85742a9_0
libwebp-base=1.2.0=h27cfd23_0
libzip=1.8.0=h4de3113_0
lz4-c=1.9.3=h9c3ff4c_1
mako=1.1.6=pypi_0
markupsafe=2.0.1=pypi_0
matplotlib=3.5.1=pypi_0
matplotlib-base=3.3.4=py39h2fa2bec_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
ncurses=6.3=h7f8727e_2
netcdf4=1.5.7=py39ha0f2276_1
numexpr=2.8.1=py39h6abb31d_0
numpy=1.21.2=py39h20f2e39_0
numpy-base=1.21.2=py39h79a1101_0
olefile=0.46=pyh9f0ad1d_1
openssl=1.1.1m=h7f8727e_0
optuna=2.10.0=pypi_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.3.5=pypi_0
patsy=0.5.2=pyhd8ed1ab_0
pbr=5.8.1=pypi_0
pillow=9.0.0=pypi_0
pip=21.2.4=py39h06a4308_0
plotly=5.5.0=pypi_0
prettytable=3.1.1=pypi_0
pymc3=3.11.2=pyh07f0a5e_1
pyparsing=3.0.6=pypi_0
pyperclip=1.8.2=pypi_0
python=3.9.7=h12debd9_1
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytz=2021.3=pyhd8ed1ab_0
pyyaml=6.0=pypi_0
readline=8.1=h27cfd23_0
scipy=1.5.3=py39hee8e79c_0
semver=2.13.0=pyh9f0ad1d_0
setuptools=58.0.4=py39h06a4308_0
six=1.16.0=pyhd3eb1b0_0
sqlalchemy=1.4.31=pypi_0
sqlite=3.37.0=hc218d9a_0
stevedore=3.5.0=pypi_0
sysroot_linux-64=2.12=he073ed8_15
tenacity=8.0.1=pypi_0
theano-pymc=1.1.2=py39h51133e4_0
tk=8.6.11=h1ccaba5_0
tornado=6.1=py39h3811e60_1
tqdm=4.62.3=pypi_0
typing-extensions=3.10.0.2=hd8ed1ab_0
typing_extensions=3.10.0.2=pyha770c72_0
tzdata=2021e=hda174b7_0
wcwidth=0.2.5=pypi_0
wheel=0.37.0=pyhd3eb1b0_1
xarray=0.21.1=pyhd8ed1ab_0
xz=5.2.5=h7b6447c_0
zipp=3.7.0=pyhd8ed1ab_1
zlib=1.2.11=h7f8727e_4
zstd=1.4.9=ha95c52a_0
